[MOD-18138] Serve queries whose union or intersection has more than 65535 children (#11238)
(cherry picked from commit 2aba0b2c35261186253eeff9d844a1c5739bbc9d)

Adapted for 8.4: this branch predates the standalone `index_result` crate, so the
capacity widening lands in its own layout, and the iterator-level tests are
replaced by an aggregate-level one where the Rust union iterator does not exist.

+++ src/redisearch_rs/low_memory_thin_vec/src/lib.rs
@@ -61,7 +61,7 @@
 //! - All Gecko-specific code has been removed
 //! - `ThinVec::drain`, `ThinVec::append` and `ThinVec::splice` have been removed, since they aren't
 //!   needed for our purposes and they contribute a significant amount of unsafe-related complexity.
-//! - Maximum capacity is limited to `u16::MAX` elements for non-zero-sized types.
+//! - Maximum capacity is limited to `u32::MAX` elements for non-zero-sized types.
 //!
 //! ## License
 //!
@@ -2066,7 +2066,7 @@ mod tests {
         }
 
         const HEADER_SIZE: usize = core::mem::size_of::<Header>();
-        assert_eq!(2 * core::mem::size_of::<u16>(), HEADER_SIZE);
+        assert_eq!(2 * core::mem::size_of::<u32>(), HEADER_SIZE);
 
         #[repr(C, align(128))]
         struct Funky<T>(T);
